% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bednets}
\alias{bednets}
\title{Bed nets}
\format{
data.frame
}
\usage{
data("bednets")
}
\description{
An example dataset with use of insecticide-treated bed nets (SH.MLR.NETS.ZS)
from UNICEF.
}
\details{
\describe{
\item{iso3c}{ISO3 country code}
\item{year}{Year}
\item{indicator}{Indicator code}
\item{value}{Data value}
\item{note}{Footnote}
\item{source}{Data source}
}
}
\keyword{datasets}
