% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_datasets.R
\name{compare_datasets}
\alias{compare_datasets}
\title{Compare datasets}
\usage{
compare_datasets(new, current, alpha = 0.05)
}
\arguments{
\item{new}{data.frame: A \code{pddcs} formatted data frame.}

\item{current}{data.frame: A \code{pddcs} formatted data frame.}

\item{alpha}{numeric: Significance level for a two-tailed test. Defaults to 0.05.}
}
\value{
A tibble with the following columns added to \code{new}:
\itemize{
\item \code{current_value}: Value in \code{current} dataset.
\item \code{current_source}: Source in \code{current} dataset.
\item \code{diff}: Absolute difference between \code{value} and \code{current_value}.
\item \code{outlier}: TRUE if the \code{diff} value is an outlier.
\item \code{p_value}: p-value for the \code{diff} value.
\item \code{n_diff}: Sum of \code{diff} by country.
\item \code{n_outlier}: Sum of \code{outlier} by country.
}
}
\description{
Compare new datasets from source with current datasets in WDI/DCS.
}
\details{
The main usage of \code{compare_datasets()} is to compare the individual
country-year values in the new source dataset with the current values in
WDI/DCS.

Comparison is done by merging the two datasets (left join on \code{new}),
calculating the absolute difference between the two \code{value} columns, and then
running outlier detection on \code{diff} column.

Users should look for \strong{both} large differences in values (\code{diff}) and large
p-values (\code{p_values}) to identify outliers or other possible unwanted changes
in the data.

In the case where a few values for a specific country are substantially
different from the current dataset in WDI/DCS they should pop out as outliers
with large p-values. On the other hand it might be the case that most or all
values for a specific country have changed. In that case it is unlikely to be
any outliers, but changes can be found by inspecting the \code{diff} and
\code{n_diff} columns.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{

\dontrun{
# Fetch indicator from source
df <- fetch_indicator("SH.MED.NUMW.P3", "who")

# Compare with WDI
dl <- compare_with_wdi(df)

# Compare new (source) and current (WDI) datasets
res <- compare_datasets(new = dl$source, current = dl$wdi)
}

}
