% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_with_wdi.R
\name{compare_with_wdi}
\alias{compare_with_wdi}
\title{Compare with WDI}
\usage{
compare_with_wdi(df)
}
\arguments{
\item{df}{data.frame: A \code{pddcs} formatted data frame. Output of
\code{fetch_indicator()}.}
}
\value{
list
}
\description{
Compare a dataset retrieved with \code{\link[=fetch_indicator]{fetch_indicator()}} with WDI.
}
\details{
\code{compare_with_wdi()} returns a list of three data frames; the original dataset
(\code{source}), the data retrieved from WDI (\code{wdi}) and the rows in the source
dataset that are not present in WDI (\code{not_in_wdi}).
}
\examples{
# Compare SH.MLR.NETS.ZS data from UNICEF with WDI
data('bednets')
dl <- compare_with_wdi(bednets)
str(dl)
}
