% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank.R
\name{create_blank}
\alias{create_blank}
\title{Create blank file}
\usage{
create_blank(indicator, iso3c = "all", year = "all", type = c("data", "meta"))
}
\arguments{
\item{indicator}{character: A CETS type indicator code.}

\item{iso3c}{character: ISO3 code. Defaults to 'all'.}

\item{year}{numeric: Year. Defaults to 'all'.}

\item{type}{character: Type, either 'data' or 'meta'.}
}
\value{
data.frame
}
\description{
Create a file with empty data or footnotes to blank pre-existing values in
DCS.
}
\examples{
# A specific indicator / country /year
df <- create_blank("SP.POP.TOTL",
  iso3c = "ALB", year = 2014
)

# Multiple indicators / countries / years
df <- create_blank(
  indicator = c("SP.POP.TOTL", "SH.MLR.NETS.ZS"),
  iso3c = c("ALB", "AGO"),
  year = 2014:2015
)

# All countries and years (data)
df <- create_blank("SP.POP.TOTL",
  type = "data"
)

# All countries and years (metadata)
df <- create_blank("SP.POP.TOTL",
  type = "meta"
)
}
