% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers}
\usage{
detect_outliers(df, var = "value", alpha = 0.05)
}
\arguments{
\item{df}{data.frame: A \code{pddcs} formatted data frame. Output of
\code{fetch_indicator()}.}

\item{var}{character: Column to use for outlier detection. Defaults to "value".}

\item{alpha}{numeric: Significance level for a two-tailed test. Defaults to 0.05.}
}
\value{
tibble
}
\description{
Detect outlier values by country.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Fetch source data
df <- fetch_indicator('SH.MED.PHYS.ZS', source = 'who')

# Detect outliers (alpha = 0.05)
df2 <- detect_outliers(df)
df2[df2$outlier,]

# Detect outliers (alpha = 0.01)
df3 <- detect_outliers(df, alpha = 0.01)
df3[df3$outlier,]

}
