% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_eurostat.R
\name{fetch_eurostat}
\alias{fetch_eurostat}
\title{Fetch data from Eurostat}
\usage{
fetch_eurostat(indicator)
}
\arguments{
\item{indicator}{character: A CETS type indicator code.}
}
\value{
data.frame
}
\description{
Fetch data from Eurostat API.
}
\keyword{internal}
