% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_indicator.R
\name{fetch_indicator}
\alias{fetch_indicator}
\title{Fetch indicator}
\usage{
fetch_indicator(indicator, source = c("eurostat", "unicef", "unpd", "who"))
}
\arguments{
\item{indicator}{character: A CETS type indicator code. See details.}

\item{source}{character: Data source.}
}
\value{
data.frame
}
\description{
Fetch data for a specified indicator from available sources.
}
\details{
See \code{?indicatorlist} for an overview of available indicators.
}
\examples{
\dontrun{
# Fetch population data from Eurostat
df <- fetch_indicator('SP.POP.TOTL', source = 'eurostat')

# Fetch population data from UNPD
df <- fetch_indicator('SP.POP.TOTL', source = 'unpd')

# Fetch bed nets data from UNICEF
df <- fetch_indicator('SH.MLR.NETS.ZS', source = 'unicef')
}
}
