% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_indicator.R
\name{fetch_indicator_single}
\alias{fetch_indicator_single}
\title{Fetch indicator (single)}
\usage{
fetch_indicator_single(
  indicator,
  source = c("eurostat", "unicef", "unpd", "who")
)
}
\arguments{
\item{indicator}{character: A CETS type indicator code. See details.}

\item{source}{character: Data source.}
}
\description{
Fetch indicator (single)
}
\keyword{internal}
