% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_unicef.R
\name{fetch_unicef}
\alias{fetch_unicef}
\title{Fetch data from UNICEF}
\usage{
fetch_unicef(indicator)
}
\arguments{
\item{indicator}{character: A CETS type indicator code.}
}
\value{
data.frame
}
\description{
Fetch data from UNICEF API.
}
\keyword{internal}
