% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_unpd.R
\name{fetch_unpd}
\alias{fetch_unpd}
\title{Fetch data from UN World Population Prospects}
\usage{
fetch_unpd(indicator)
}
\arguments{
\item{indicator}{character: A CETS type indicator code.}
}
\value{
data.frame
}
\description{
Fetch data from the UN Population Division's World Population Prospects.
}
\keyword{internal}
