% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_wdi.R
\name{fetch_wdi}
\alias{fetch_wdi}
\title{Fetch data from WDI}
\usage{
fetch_wdi(indicator)
}
\arguments{
\item{indicator}{character: A WDI indicator code.}
}
\value{
data.frame
}
\description{
Fetch data from WDI API.
}
\keyword{internal}
