% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_who.R
\name{fetch_who}
\alias{fetch_who}
\title{Fetch data from WHO}
\usage{
fetch_who(indicator)
}
\arguments{
\item{indicator}{character: A CETS type indicator code.}
}
\value{
data.frame
}
\description{
Fetch data from WHO API.
}
\keyword{internal}
