% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorlist}
\alias{indicatorlist}
\title{Supported indicators}
\format{
data.frame
}
\usage{
indicatorlist
}
\description{
A dataset containing the available indicators for \code{\link[=fetch_indicator]{fetch_indicator()}}.
}
\details{
List of available indicators, by data source.

Eurostat:
\itemize{
\item SP.DYN.CBRT.IN
\item SP.POP.TOTL
}

UNPD:
\itemize{
\item SP.DYN.TFRT.IN
\item SP.POP.AG00.MA.IN - SP.POP.AG25.MA.IN
\item SP.POP.TOTL
}

UNICEF:
\itemize{
\item SH.MLR.NETS.ZS
\item SH.STA.ANV4.ZS
\item SH.STA.ORTH
\item SN.ITK.VITA.ZS
}

WHO:
\itemize{
\item SH.ALC.PCAP.LI, SH.ALC.PCAP.MA.LI, SH.ALC.PCAP.FE.LI
\item SH.ANM.ALLW.ZS
\item SH.ANM.CHLD.ZS
\item SH.ANM.NPRG.ZS
\item SH.DYN.NCOM.ZS, SH.DYN.NCOM.MA.ZS, SH.DYN.NCOM.FE.ZS
\item SH.MED.BEDS.ZS
\item SH.MED.NUMW.P3
\item SH.MED.PHYS.ZS
\item SH.PRG.ANEM
\item SH.PRG.SYPH.ZS
\item SH.PRV.SMOK, SH.PRV.SMOK.MA, SH.PRV.SMOK.FE
\item SH.STA.AIRP.P5, SH.STA.AIRP.MA.P5, SH.STA.AIRP.FE.P5
\item SH.STA.OWAD.ZS, SH.STA.OWAD.MA.ZS, SH.STA.OWAD.FE.ZS
\item SH.STA.POIS.P5, SH.STA.POIS.P5.MA, SH.STA.POIS.P5.FE
\item SH.STA.SUIC.P5, SH.STA.SUIC.MA.P5, SH.STA.SUIC.FE.P5
\item SH.STA.TRAF.P5
\item SH.STA.WASH.P5
\item SP.REG.DTHS.ZS
}
}
\keyword{datasets}
