% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Population}
\format{
data.frame
}
\usage{
data("population")
}
\description{
An example dataset with population data (SP.POP.TOTL) from Eurostat.
}
\details{
\describe{
\item{iso3c}{ISO3 country code}
\item{year}{Year}
\item{indicator}{Indicator code}
\item{value}{Data value}
\item{note}{Footnote}
\item{source}{Data source}
}
}
\keyword{datasets}
