% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data}
\usage{
prepare_data(path, source)
}
\arguments{
\item{path}{character: Path to file.}

\item{source}{character: Data source.}
}
\value{
tibble
}
\description{
Prepare internal data for DCS upload.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Prepare UNODC homicide data
path <- system.file("extdata",
  "homicide_all.xlsx", package = "pddcs")
df <- prepare_data(path, source = 'homicide')
}
