% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_poverty_mdim.R
\name{prepare_poverty_mdim}
\alias{prepare_poverty_mdim}
\title{Prepare poverty MDIM}
\usage{
prepare_poverty_mdim(path)
}
\arguments{
\item{path}{character: Path to file.}
}
\description{
Prepare MDIM poverty data for DCS upload.
}
\examples{
\dontrun{
# Prepare data
df <- prepare_poverty_mdim("wdi_latest.xltm")
}
}
\keyword{internal}
