% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_poverty_rmt.R
\name{prepare_poverty_rmt}
\alias{prepare_poverty_rmt}
\title{Prepare poverty RMT}
\usage{
prepare_poverty_rmt(path)
}
\arguments{
\item{path}{character: Path to file.}
}
\description{
Prepare SI.RMT.COST data for DCS upload.
}
\examples{
\dontrun{
# Prepare data
df <- prepare_poverty_rmt('UN SDG - 10.c_inquiry response.xlsx')
}
}
\keyword{internal}
