% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_unodc_homicide.R
\name{prepare_unodc_homicide}
\alias{prepare_unodc_homicide}
\title{Prepare UNODC homicide}
\usage{
prepare_unodc_homicide(path)
}
\arguments{
\item{path}{character: Path to file.}
}
\description{
Prepare UNODC homicide data for DCS upload.
}
\examples{
# Prepare data
df <- prepare_unodc_homicide('homicide_all.xls')
}
\keyword{internal}
