% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dcs.R
\name{read_dcs}
\alias{read_dcs}
\title{Read DCS}
\usage{
read_dcs(path)
}
\arguments{
\item{path}{character: Path to file.}
}
\value{
tibble
}
\description{
Read data from DCS.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Data from DCS needs to be download with Export > Custom > OK. See example
dataset for the accepted format.
}
\examples{
path <- system.file("extdata",
  "data-nurses-and-midwives-in-dcs.xlsx",
   package = "pddcs")
df <- read_dcs(path)
}
