% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_unicef.R
\name{recode_unicef_codes}
\alias{recode_unicef_codes}
\title{recode_unicef_codes}
\usage{
recode_unicef_codes(indicator)
}
\arguments{
\item{indicator}{character: A CETS type indicator code.}
}
\value{
character
}
\description{
recode_unicef_codes
}
