% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dcs.R
\name{write_dcs}
\alias{write_dcs}
\title{Write DCS}
\usage{
write_dcs(df, path, type = c("data", "meta"))
}
\arguments{
\item{df}{data.frame: A DCS formatted data frame. Output of \code{format_dcs()}.}

\item{path}{character: A file name to write to.}

\item{type}{character: Type, either 'data' or 'meta'.}
}
\description{
Write a DCS formatted dataset to a file that can be uploaded to the Data
Collection System.
}
\examples{
\dontrun{
# Load example data
data("bednets")

# Write DCS 'data' format
df <- format_dcs(bednets, type = 'data')
write_dcs(df, path = 'data-bednets.xlsx', type = 'data')

# Write DCS 'metadata' format
df <- format_dcs(bednets, type = 'meta')
write_dcs(df, path = 'meta-bednets.xlsx', type = 'meta')
}
}
